♻️ Make u128

+++ src/libraries/SuperConversionLib.sol
@@ -3,103 +3,103 @@ pragma solidity ^0.8.0;
 
 /// @author philogy <https://github.com/philogy>
 library SuperConversionLib {
-    function into(uint256[1] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[1] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[2] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[2] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[3] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[3] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[4] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[4] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[5] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[5] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[6] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[6] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[7] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[7] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[8] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[8] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[9] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[9] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[10] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[10] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[11] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[11] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[12] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[12] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[13] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[13] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[14] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[14] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[15] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[15] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[16] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[16] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[17] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[17] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[18] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[18] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[19] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[19] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[20] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[20] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[21] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[21] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[22] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[22] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[23] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[23] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[24] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[24] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 
-    function into(uint256[25] memory arr) internal pure returns (uint256[] memory) {
-        return abi.decode(abi.encode(uint256(0x20), arr.length, arr), (uint256[]));
+    function into(uint128[25] memory arr) internal pure returns (uint128[] memory) {
+        return abi.decode(abi.encode(uint128(0x20), arr.length, arr), (uint128[]));
     }
 }
